Reorganized structure to play nicely with tests Instead of manually having to handle our server connection for tests, jest now takes care of it for us.
I tried to solve the openHandles issue jest was giving me,
could not find where the problem lies! :(

By having jest handle our port connection, app.listen() is
no longer the possbile culprit, --detectOpenHandles seems
to point to one of supertest's methods but no change in
syntax seems to resolve the issue, so it might just be
within the library itself. Out of our control.

I had also tried manually closing our server and DB connection,
however that did not change the openHandles issue.A

Still, I made some nice changes to the structure that I still
want to commit and issue #48 should just be left as unresolved
since it is not imperative to the app AND the issue itself was
technically resolved.

Might make a separate issue for Jest, just to keep in the back
of my mind.